fixed home page category section
wanted to make it so that. instead of a button the actual image was clickable. which proved to be a fair challenge in itself.

however i managed to do this by removing the div tags to the shop button and anchoring them instead to the image container containing the image.